chore(trace): move writer module
