+++ ddtrace/_trace/writer/_encoding.pyi
