+++ ddtrace/_trace/writer/_encoding.pyx
