+++ ddtrace/_trace/writer/encoding.py
