+++ ddtrace/profiling/profiler.py
@@ -10,12 +10,12 @@ from typing import Union  # noqa:F401
 import attr
 
 import ddtrace
+from ddtrace._trace import writer
 from ddtrace.internal import agent
 from ddtrace.internal import atexit
 from ddtrace.internal import forksafe
 from ddtrace.internal import service
 from ddtrace.internal import uwsgi
-from ddtrace.internal import writer
 from ddtrace.internal.datadog.profiling import ddup
 from ddtrace.internal.module import ModuleWatchdog
 from ddtrace.profiling import collector
