+++ pyproject.toml
@@ -84,7 +84,7 @@ exclude = '''
   .venv*
   | \.riot/
   | ddtrace/appsec/_ddwaf.pyx$
-  | ddtrace/internal/_encoding.pyx$
+  | ddtrace/_trace/writer/_encoding.pyx$
   | ddtrace/internal/_rand.pyx$
   | ddtrace/internal/_tagset.pyx$
   | ddtrace/profiling/collector/_traceback.pyx$
