+++ tests/integration/test_encoding.py
@@ -13,7 +13,7 @@ AGENT_VERSION = os.environ.get("AGENT_VERSION")
 class TestTraceAcceptedByAgent:
     def test_simple_trace_accepted_by_agent(self):
         tracer = Tracer()
-        with mock.patch("ddtrace.internal.writer.writer.log") as log:
+        with mock.patch("ddtrace._trace.writer.log") as log:
             with tracer.trace("root"):
                 for _ in range(999):
                     with tracer.trace("child"):
@@ -33,7 +33,7 @@ class TestTraceAcceptedByAgent:
     def test_trace_with_meta_accepted_by_agent(self, tags):
         """Meta tags should be text types."""
         tracer = Tracer()
-        with mock.patch("ddtrace.internal.writer.writer.log") as log:
+        with mock.patch("ddtrace._trace.writer.log") as log:
             with tracer.trace("root", service="test_encoding", resource="test_resource") as root:
                 root.set_tags(tags)
                 for _ in range(999):
@@ -54,7 +54,7 @@ class TestTraceAcceptedByAgent:
     def test_trace_with_metrics_accepted_by_agent(self, metrics):
         """Metric tags should be numeric types - i.e. int, float, long (py3), and str numbers."""
         tracer = Tracer()
-        with mock.patch("ddtrace.internal.writer.writer.log") as log:
+        with mock.patch("ddtrace._trace.writer.log") as log:
             with tracer.trace("root") as root:
                 root.set_metrics(metrics)
                 for _ in range(999):
